Move report_region_errors to region_errors.rs
